Rename tagging dir to checksum
