+++ lib/src/checksum/hash.spec.ts
