+++ lib/src/checksum/hash.ts
